requirements.txt
aiosqlite==0.20.0
APScheduler==3.10.4
fastapi==0.111.0
httpx==0.27.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
uvicorn==0.30.1
flake8==7.1.0